Show that setting h0_snow has no effect in snow components
The initial snowpack depth, `h0_snow`, is a parameter in the cfg
files for both SnowDegreeDay and SnowEnergyBalance. As its name
implies, it's the first value of snowpack depth, `h_snow`. However,
its value doesn't matter, since in the first update of either
component, the `h_snow` value is recalculated from SWE and the
densities of water and snow.